backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
email-validator==2.1.1
boto3==1.34.34
numpy==1.26.4
ezdxf==1.2.0
redis==5.0.1
rq==1.15.1
